Refine code
